Add Hybrid Inflation to Shiden (#1095)
* Added hybrid inflation to Shiden + Migration

* updated RewardDistributionConfig
